treewide: Remove oplus secure guard
Well this includes a lot of tracing and dirty hacks, e.g. hijack the
critical syscall path just to identify processes that is trying to
setuid() and kill them before getting root permission. They gracefully
name it "root guard" but it is stupid enough that people would laugh at.

Although not all of the features are enabled, totally remove them from
the tree.

Signed-off-by: LibXZR <i@xzr.moe>